Fixed a few bugs, and added more features

+++ js/user/deleteUser.js
@@ -1,16 +1,19 @@
-$(document).ajaxComplete(function() {
+$(document).ajaxStop(function() {
   $('form.delete').submit(function(event) {
-    var data = $(this).serializeArray();
-    $.ajax({
-      url: domain + "user/" + data[0].value,
-      type: "DELETE",
-      success: function (result, status, xhr) {
-        window.location = "/cedar/user/";
-      },
-      error: function(xhr, status, error) {
-        console.log(xhr.statusText);
-      },
-    });
+    var d = confirm("Are you sure you want to delete the user?");
+    if (d) {
+      var data = $(this).serializeArray();
+      $.ajax({
+        url: domain + "user/" + data[0].value,
+        type: "DELETE",
+        success: function (result, status, xhr) {
+          window.location = "/cedar/user/";
+        },
+        error: function(xhr, status, error) {
+          console.log(xhr.statusText);
+        },
+      });
+    }
     event.preventDefault();
   });
 });
